Refactor ScheduleOverlap.vue into mixins and child components
Decompose the 4,625-line ScheduleOverlap.vue into smaller, focused modules
to improve maintainability and readability:

- Extract dragMixin.js (425 lines): drag interaction, touch/mouse events,
  coordinate math, column offsets
- Extract availabilityMixin.js (814 lines): response fetching, availability
  parsing, calendar event integration, submit/delete flows, watchers
- Extract signUpMixin.js (183 lines): sign-up block CRUD, form reset,
  submission logic
- Extract DaysOnlyCalendar.vue (108 lines): monthly calendar grid template
  with pagination and hint text
- Extract EditAvailabilitySidebar.vue (328 lines): edit-availability sidebar
  with calendar options, overlay toggle, guest name editing, delete dialog

ScheduleOverlap.vue is reduced to 2,894 lines (37% reduction). All $refs
calls from Event.vue and LandingPageCalendar.vue remain functional since
mixin methods merge onto `this`.
